requirements.txt
python-rrmngmnt